requirement.txt
cffi==1.15.0
click==8.1.2
colorama==0.4.4
cryptography==37.0.2
distlib==0.3.4
filelock==3.6.0
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.29
platformdirs==2.5.1
protobuf==3.20.1
pycparser==2.21
PyMySQL==1.0.2
six==1.16.0
SQLAlchemy==1.4.35
virtualenv==20.14.0
Werkzeug==2.1.1
